requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain-google-genai==0.0.19
chromadb==0.5.3
python-dotenv==1.0.1
pydantic==2.8.2
langchain-text-splitters==0.2.2
langchain-core==0.2.16